safari-600.7-branch/LayoutTests/webarchive/test-link-href.html

<html>
<head>
<script>
    if (window.testRunner)
        testRunner.dumpDOMAsWebArchive();
</script>
<link rel="stylesheet" href="resources/test-style.css" type="text/css">
</head>
<body>
<div>
This text should be red sans-serif.
</div>
</body>
</html>
